requirements.txt
fake-factory==0.5.0
psycopg2==2.5.4
requests==2.7.0
--allow-external salesforce-python-toolkit
--allow-unverified salesforce-python-toolkit
salesforce-python-toolkit==0.1.5
SQLAlchemy==0.9.8
suds==0.3.9
